add response check for status check

+++ src/store/bid/moduleBidActions.js
@@ -16,7 +16,7 @@ export default {
     } catch (err) {
       if (state.apiCounter === 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getTeamMembers', payload);
@@ -35,7 +35,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getSalesReps', payload);
@@ -54,7 +54,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('searchByCompany', payload);
@@ -74,7 +74,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getCompanyByServices', payload);
@@ -91,7 +91,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getDraftBids', payload);
@@ -109,7 +109,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getBidsLists', payload);
@@ -150,7 +150,7 @@ export default {
       commit('setViewBidError', true);
       if (state.bidApiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.bidApiCounter = 2;
         dispatch('getBidBySerial', payload);
@@ -173,7 +173,7 @@ export default {
       commit('setErrorDeleteBid');
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('deleteBid', payload);
@@ -202,7 +202,7 @@ export default {
       commit('setPageLoader', false);
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getIntent', payload);
@@ -227,7 +227,7 @@ export default {
       commit('setPageLoader', false);
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getAllIntent', payload);
@@ -247,7 +247,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('updateIntent', payload);
@@ -271,7 +271,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('makeIntent', payload);
@@ -297,7 +297,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('awardCompany', payload);
@@ -322,7 +322,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('rejectCompany', payload);
@@ -347,7 +347,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('UnAwardCompany', payload);
@@ -371,7 +371,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('UnDisqualifyCompany', payload);
@@ -428,7 +428,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('submitBid', payload);
@@ -494,7 +494,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('editSubmitBid', payload);
@@ -531,7 +531,7 @@ export default {
       commit('setPageLoader', false);
       if (state.apiCounter === 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getQA', payload);
@@ -554,7 +554,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('askQuestion', payload);
@@ -577,7 +577,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('answerQuestion', payload);
@@ -598,7 +598,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('editAnswer', payload);
@@ -718,7 +718,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('saveDraftBid', payload);
@@ -835,7 +835,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('updateDraftBid', payload);
@@ -856,7 +856,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('inviteNewSupplier', payload);
@@ -933,7 +933,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('uploadBidAttach', payload);
@@ -949,7 +949,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getBidTemplates', payload);
@@ -997,7 +997,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('deleteTemplate', payload);
@@ -1013,7 +1013,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('deleteTemplate', payload);
@@ -1040,7 +1040,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getEditTemplate', payload);
@@ -1098,7 +1098,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('saveTemplateBid', payload);
@@ -1126,7 +1126,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getBidBySerial', payload);
@@ -1323,7 +1323,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
 
@@ -1379,7 +1379,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('deleteDraftBid', payload);
@@ -1413,7 +1413,7 @@ export default {
       commit('setPageLoader', false);
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('getUpdateBid', payload);
@@ -1539,7 +1539,7 @@ export default {
     } catch (err) {
       if (state.apiCounter == 2) {
         dispatch('apiSignOutAction');
-      } else if (err.response.status === 403) {
+      } else if (err.response && err.response.status === 403) {
         await dispatch('refreshToken');
         state.apiCounter = 2;
         dispatch('updateBid', payload);
